backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic[email]>=2.0
pydantic-settings
firebase-admin>=6.0
python-jose[cryptography]
passlib[bcrypt]
langchain
langchain-openai
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
httpx
hypothesis
pytest
pytest-asyncio
